Move importer tests to dedicated folder (#380)
