+++ features/facebookImport/test/importers/ad-views-importer.test.js
@@ -1,6 +1,6 @@
 "use strict";
 
-import { RECENTLY_VIEWED_FILE_PATH } from "../src/model/importers/recently-viewed-ads-importer";
+import { RECENTLY_VIEWED_FILE_PATH } from "../../src/model/importers/recently-viewed-ads-importer";
 import {
     creatAdViewsWithCompanyWithUnicodeCharactersData,
     createDanishAdViewsData,
@@ -8,17 +8,17 @@ import {
     createEnglishDatasetWithMissingAdsCategory,
     createGermanAdViewsData,
     createIncompleteEnglishAdViewsData,
-} from "./datasets/ad-views-data";
-import { ZipFileMock } from "./mocks/zipfile-mock";
+} from "../datasets/ad-views-data";
+import { ZipFileMock } from "../mocks/zipfile-mock";
 import {
     runAdsImportForDataset,
     runRecentlyViewedAdsImporter,
-} from "./utils/data-importing";
+} from "../utils/data-importing";
 import {
     expectImportSuccess,
     expectInvalidContentError,
     expectMissingFileError,
-} from "./utils/importer-assertions";
+} from "../utils/importer-assertions";
 
 describe("Import ad views from empty export", () => {
     let zipFile = null;
