+++ features/facebookImport/test/importers/messages-importer.test.js
@@ -1,12 +1,12 @@
 "use strict";
 
-import { MissingMessagesFilesException } from "../src/model/importers/utils/failed-import-exception.js";
-import { ZipFileMock } from "./mocks/zipfile-mock.js";
-import { runMessagesImporter } from "./utils/data-importing.js";
+import { MissingMessagesFilesException } from "../../src/model/importers/utils/failed-import-exception.js";
+import { ZipFileMock } from "../mocks/zipfile-mock.js";
+import { runMessagesImporter } from "../utils/data-importing.js";
 import {
     expectError,
     expectImportSuccess,
-} from "./utils/importer-assertions.js";
+} from "../utils/importer-assertions.js";
 
 function createMessageThreadOneData() {
     return {
