+++ features/facebookImport/test/importers/off-facebook-events-importer.test.js
@@ -1,14 +1,14 @@
 "use strict";
 
-import { OFF_FACEBOOK_EVENTS_FILE_PATH } from "../src/model/importers/off-facebook-events-importer";
-import { createOffFacebookEventsSimpleData } from "./datasets/off-facebook-events-data";
-import { ZipFileMock } from "./mocks/zipfile-mock";
-import { runOffFacebookEventsImporter } from "./utils/data-importing";
+import { OFF_FACEBOOK_EVENTS_FILE_PATH } from "../../src/model/importers/off-facebook-events-importer";
+import { createOffFacebookEventsSimpleData } from "../datasets/off-facebook-events-data";
+import { ZipFileMock } from "../mocks/zipfile-mock";
+import { runOffFacebookEventsImporter } from "../utils/data-importing";
 import {
     expectImportSuccess,
     expectInvalidContentError,
     expectMissingFileError,
-} from "./utils/importer-assertions";
+} from "../utils/importer-assertions";
 
 describe("Import off-facebook events from empty export", () => {
     let zipFile = null;
